Move editor config to its own file
